Make C++ API private
'include' directory now contains the C API only
