+++ libs/Microsoft.MixedReality.WebRTC.Native/src/audio_frame_observer.h
