+++ libs/Microsoft.MixedReality.WebRTC.Native/src/interop/external_video_track_source_interop.cpp
@@ -6,8 +6,8 @@
 #include "pch.h"
 
 #include "callback.h"
-#include "external_video_track_source.h"
-#include "interop/external_video_track_source_interop.h"
+#include "media/external_video_track_source.h"
+#include "external_video_track_source_interop.h"
 
 using namespace Microsoft::MixedReality::WebRTC;
