+++ libs/Microsoft.MixedReality.WebRTC.Native/src/interop/local_video_track_interop.cpp
@@ -5,8 +5,8 @@
 // line, to prevent clang-format from reordering it with other headers.
 #include "pch.h"
 
-#include "interop/local_video_track_interop.h"
-#include "local_video_track.h"
+#include "local_video_track_interop.h"
+#include "media/local_video_track.h"
 
 using namespace Microsoft::MixedReality::WebRTC;
