+++ libs/Microsoft.MixedReality.WebRTC.Native/src/mrs_errors.h
@@ -6,76 +6,9 @@
 #include <string>
 
 #include "export.h"
+#include "result.h"
 
 namespace Microsoft::MixedReality::WebRTC {
-
-/// Result code from an operation, typically used through the interop layer
-/// instead of a full-featured Error object.
-///
-/// Somewhat similar to webrtc::RTCErrorType to avoid pulling it as a dependency in the
-/// public API. This also has extra values not found in webrtc::RTCErrorType.
-enum class Result : std::uint32_t {
-  /// The operation was successful.
-  kSuccess = 0,
-
-  //
-  // Generic errors
-  //
-
-  /// Unknown internal error.
-  /// This is generally the fallback value when no other error code applies.
-  kUnknownError = 0x80000000,
-
-  /// A parameter passed to the API function was invalid.
-  kInvalidParameter = 0x80000001,
-
-  /// The operation cannot be performed in the current state.
-  kInvalidOperation = 0x80000002,
-
-  /// A call was made to an API function on the wrong thread.
-  /// This is generally related to platforms with thread affinity like UWP.
-  kWrongThread = 0x80000003,
-
-  /// An object was not found.
-  kNotFound = 0x80000004,
-
-  /// An interop handle referencing a native object instance is invalid,
-  /// although the API function was expecting a valid object.
-  kInvalidNativeHandle = 0x80000005,
-
-  /// The API object is not initialized, and cannot as a result perform the
-  /// given operation.
-  kNotInitialized = 0x80000006,
-
-  /// The current operation is not supported by the implementation.
-  kUnsupported = 0x80000007,
-
-  /// An argument was passed to the API function with a value out of the expected range.
-  kOutOfRange = 0x80000008,
-
-  //
-  // Peer connection (0x1xx)
-  //
-
-  /// The peer connection is closed, but the current operation requires an open
-  /// peer connection.
-  kPeerConnectionClosed = 0x80000101,
-
-  //
-  // Data (0x3xx)
-  //
-
-  /// The SCTP handshake for data channels encryption was not performed, because
-  /// the connection was established before any data channel was added to it.
-  /// Due to limitations in the implementation, without SCTP handshake data
-  /// channels cannot be used, and therefor applications expecting to use data
-  /// channels must open at least a single channel before establishing a peer
-  /// connection (calling |CreateOffer()|).
-  kSctpNotNegotiated = 0x80000301,
-
-  /// The specified data channel ID is invalid.
-  kInvalidDataChannelId = 0x80000302,
-};
 
 /// Full-featured error object, containing an error code and a message.
 ///
@@ -85,7 +18,7 @@ class Error {
  public:
   /// Create an empty "error" wrapping a non-error result.
   /// Preferred over the default constructor for readability.
-  static MRS_API Error OK();
+  static Error OK();
 
   /// Create an empty "error" wrapping a non-error result.
   Error() = default;
@@ -103,19 +36,19 @@ class Error {
   Error(const Error& other) = delete;
   Error& operator=(const Error& other) = delete;
 
-  MRS_API Error(Error&& other);
-  MRS_API Error& operator=(Error&& other);
+  Error(Error&& other);
+  Error& operator=(Error&& other);
 
   Result result() const { return result_; }
   void set_result(Result result) { result_ = result; }
 
   /// Human-readable informational message, for display only.
   /// The message is susceptible to change in future revisions.
-  MRS_API const char* message() const;
+  const char* message() const;
 
   /// Explicitly set the informational message associated with the error, often
   /// to provide more context than the default one.
-  MRS_API void set_message(std::string message);
+  void set_message(std::string message);
 
   /// Return |true| if the Error instance does not currently represent an error.
   bool ok() const { return result_ == Result::kSuccess; }
