+++ libs/Microsoft.MixedReality.WebRTC.Native/src/tracked_object.h
