+++ libs/Microsoft.MixedReality.WebRTC.Native/src/uwp/Microsoft.MixedReality.WebRTC.Native.UWP.vcxproj.filters
@@ -1,97 +1,63 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="../audio_frame_observer.cpp">
-      <Filter>media</Filter>
-    </ClCompile>
     <ClCompile Include="../pch.cpp" />
-    <ClCompile Include="../peer_connection.cpp" />
-    <ClCompile Include="../video_frame_observer.cpp">
-      <Filter>media</Filter>
-    </ClCompile>
-    <ClCompile Include="../sdp_utils.cpp" />
-    <ClCompile Include="../interop/interop_api.cpp">
-      <Filter>interop</Filter>
-    </ClCompile>
-    <ClCompile Include="../data_channel.cpp">
-      <Filter>media</Filter>
-    </ClCompile>
-    <ClCompile Include="../str.cpp">
-      <Filter>interop</Filter>
-    </ClCompile>
-    <ClCompile Include="../interop/peer_connection_interop.cpp">
-      <Filter>interop</Filter>
-    </ClCompile>
-    <ClCompile Include="../interop/global_factory.cpp">
-      <Filter>interop</Filter>
-    </ClCompile>
-    <ClCompile Include="../media/local_video_track.cpp">
-      <Filter>media</Filter>
-    </ClCompile>
-    <ClCompile Include="../interop/local_video_track_interop.cpp">
-      <Filter>interop</Filter>
-    </ClCompile>
+    <ClCompile Include="..\audio_frame_observer.cpp" />
+    <ClCompile Include="..\data_channel.cpp" />
     <ClCompile Include="..\mrs_errors.cpp" />
-    <ClCompile Include="../interop/external_video_track_source_interop.cpp">
+    <ClCompile Include="..\peer_connection.cpp" />
+    <ClCompile Include="..\sdp_utils.cpp" />
+    <ClCompile Include="..\str.cpp" />
+    <ClCompile Include="..\video_frame_observer.cpp" />
+    <ClCompile Include="..\interop\external_video_track_source_interop.cpp">
+      <Filter>interop</Filter>
+    </ClCompile>
+    <ClCompile Include="..\interop\global_factory.cpp">
+      <Filter>interop</Filter>
+    </ClCompile>
+    <ClCompile Include="..\interop\interop_api.cpp">
+      <Filter>interop</Filter>
+    </ClCompile>
+    <ClCompile Include="..\interop\local_video_track_interop.cpp">
+      <Filter>interop</Filter>
+    </ClCompile>
+    <ClCompile Include="..\interop\peer_connection_interop.cpp">
       <Filter>interop</Filter>
     </ClCompile>
     <ClCompile Include="..\media\external_video_track_source.cpp">
       <Filter>media</Filter>
     </ClCompile>
+    <ClCompile Include="..\media\local_video_track.cpp">
+      <Filter>media</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="../../include/audio_frame_observer.h">
-      <Filter>media</Filter>
-    </ClInclude>
-    <ClInclude Include="../../include/callback.h" />
     <ClInclude Include="../pch.h" />
-    <ClInclude Include="../../include/peer_connection.h" />
-    <ClInclude Include="../targetver.h" />
-    <ClInclude Include="../../include/video_frame_observer.h">
-      <Filter>media</Filter>
-    </ClInclude>
-    <ClInclude Include="../../include/sdp_utils.h" />
-    <ClInclude Include="../interop/interop_api.h">
-      <Filter>interop</Filter>
-    </ClInclude>
-    <ClInclude Include="../../include/str.h">
-      <Filter>interop</Filter>
-    </ClInclude>
-    <ClInclude Include="../../include/export.h" />
-    <ClInclude Include="../interop/peer_connection_interop.h">
-      <Filter>interop</Filter>
-    </ClInclude>
-    <ClInclude Include="../interop/global_factory.h">
-      <Filter>interop</Filter>
-    </ClInclude>
-    <ClInclude Include="../../include/local_video_track.h">
-      <Filter>media</Filter>
-    </ClInclude>
-    <ClInclude Include="../../include/data_channel.h">
-      <Filter>media</Filter>
-    </ClInclude>
-    <ClInclude Include="../interop/local_video_track_interop.h">
-      <Filter>interop</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\include\tracked_object.h" />
-    <ClInclude Include="..\..\include\ref_counted_base.h" />
-    <ClInclude Include="..\..\include\refptr.h" />
-    <ClInclude Include="..\..\include\mrs_errors.h" />
-    <ClInclude Include="..\..\include\audio_frame.h">
-      <Filter>media</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\include\video_frame.h">
-      <Filter>media</Filter>
-    </ClInclude>
-    <ClInclude Include="../interop/external_video_track_source_interop.h">
+    <ClInclude Include="..\audio_frame_observer.h" />
+    <ClInclude Include="..\callback.h" />
+    <ClInclude Include="..\data_channel.h" />
+    <ClInclude Include="..\mrs_errors.h" />
+    <ClInclude Include="..\peer_connection.h" />
+    <ClInclude Include="..\ref_counted_base.h" />
+    <ClInclude Include="..\refptr.h" />
+    <ClInclude Include="..\sdp_utils.h" />
+    <ClInclude Include="..\str.h" />
+    <ClInclude Include="..\targetver.h" />
+    <ClInclude Include="..\tracked_object.h" />
+    <ClInclude Include="..\video_frame_observer.h" />
+    <ClInclude Include="..\interop\global_factory.h">
       <Filter>interop</Filter>
     </ClInclude>
     <ClInclude Include="..\media\external_video_track_source_impl.h">
       <Filter>media</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\include\external_video_track_source.h">
+    <ClInclude Include="..\media\external_video_track_source.h">
       <Filter>media</Filter>
     </ClInclude>
+    <ClInclude Include="..\media\local_video_track.h">
+      <Filter>media</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\include\result.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="../../docs/design.md" />
@@ -99,10 +65,10 @@
   </ItemGroup>
   <ItemGroup>
     <Filter Include="interop">
-      <UniqueIdentifier>{3313d587-9205-4a7e-989b-699a8d91d7d6}</UniqueIdentifier>
+      <UniqueIdentifier>{2c834693-8aec-4ee1-9c56-b8827db37535}</UniqueIdentifier>
     </Filter>
     <Filter Include="media">
-      <UniqueIdentifier>{f75c7170-eb40-43f9-b14a-45a8ef81b975}</UniqueIdentifier>
+      <UniqueIdentifier>{3716966c-d0c2-4ecc-a361-12fea77b532b}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
 </Project>
