+++ libs/Microsoft.MixedReality.WebRTC.Native/src/video_frame_observer.h
@@ -114,10 +114,10 @@ class VideoFrameObserver : public rtc::VideoSinkInterface<webrtc::VideoFrame> {
   /// be reused by a later call so concurrent access is not supported.
   /// Before calling this method the caller needs to acquire |mutex_|, and keep
   /// it for the duration of its access.
-  MRS_API ArgbBuffer* GetArgbScratchBuffer(int width, int height);
+  ArgbBuffer* GetArgbScratchBuffer(int width, int height);
 
   // VideoSinkInterface interface
-  MRS_API void OnFrame(const webrtc::VideoFrame& frame) noexcept override;
+  void OnFrame(const webrtc::VideoFrame& frame) noexcept override;
 
  private:
   /// Registered callback for receiving I420-encoded frame.
